Fix dangerous ownership
This patch fixes a common mistake with manually installed software, which adds the software, in a system-wide installation path, but leave it owned to the user. If an admin start to execute this binary, it's wonderful privilege escalation vector.

This patch resolves the problem by chowning it first, and therefore require root privileges to manipulate the file, resolving the potential attack vector.