Update NavigationDocs.jsx to be more clear about installing the NetBird client
I've seen several people get very confused about setting up NetBird in terms of self-hosting when this is the first thing they see on the docs. Reminds me of the problem with 'Plex Media Server' versus 'Plex' with a lot of people installing the server on their computer instead of the client software.