Moved Self-hosted vs. Cloud-hosted NetBird in navigation menu
It makes far more sense to have this under `Use Cases`, IMO.

It's the kind of information a user would like to have upfront and not as an afterthought. Furthermore, comparisons between the two options are typically made far sooner in other projects.

May also consider changing the dir that the `.md` file is in, which also change the URL to `/selfhosted/` from to possibly `how-to`, though that's probably beyond the scope of what I can pull off from my cellphone in bed. :- ) Edits of course welcome re that.